Branch: Kris-AR * src/conf/modules/filters.options    Added reference to new mid fetch filter that will abort the download    of HTTP documents that have not changed according to their    last-modified and etags header values * src/java/is/hi/bok/crawler/ar/frontier/AdaptiveRevisitFrontier.java    Synchronized access to two methods. They access data that requires    exclusive access. * src/java/is/hi/bok/crawler/ar/filter/ChangeEvaluator.java    Now set's content size to zero for documents that have    not changed. This keeps the StatisticsTracker from including    downloaded duplicates in the tally of 'written' data. * src/webapps/admin/console/readalert.jsp    Fix for NPE that occured if a non-existing job was    selected.

+++ heritrix/src/webapps/admin/console/readalert.jsp
@@ -6,16 +6,18 @@
 
 <%
     Alert alert = Heritrix.getAlert(request.getParameter("alert"));
-    alert.setAlertSeen();
     String title = "Read alert";
     int tab = 0;
 %>
 
 <%@include file="/include/head.jsp"%>
 <p>
-<% if(alert == null){ %>
-    <b> No matching alert found </b>
-<% } else { %>
+<%  if(alert == null){  %>
+        <b> No matching alert found </b>
+<% 
+    } else { 
+        alert.setAlertSeen();
+%>
     <table>
         <tr>
             <td>
@@ -54,5 +56,7 @@
     <p>
         <a href="<%=request.getContextPath()%>/console/alerts.jsp">Back to alerts</a>
         &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
-        <a href="<%=request.getContextPath()%>/console/alerts.jsp?alerts=<%=alert.getID()%>&action=delete">Delete this alert</a>
+        <% if(alert != null) { %>
+            <a href="<%=request.getContextPath()%>/console/alerts.jsp?alerts=<%=alert.getID()%>&action=delete">Delete this alert</a>
+        <% } %>
 <%@include file="/include/foot.jsp"%>
